Merge modules into top level declarations with the same name whenever it's possible
This contributes to #58.
This is only the first step for solving the issue.
In fact I'm pretty much aware that at least one unit test, escaping.d.ts
is actually wrong - it's not changed because I don't have a decent
solution so far and at least I know that current (erroneous in some
aspects) behavior is consistent with behaviour of code before.

As soon as I'll figure it out, mergeClassLikesAndModuleDeclarations can
be removed because mergeWithNameSpace is actaully the lowering that does
things right.

see DUKAT-2