Update nodejs version to 13.5.0
This version shows significant improvements in terms of memory
management and we need this for tests. This does not mean, however,
that this is the minimal version. In fact the js part is intentionally
written that way so that we are not relying on any modern features
(and won't in forseeable future).