ModelLowering should implement fun lower(module: ModuleModel): ModuleModel
Sometimes this leads to introducing lateinit var pattern which I'm not the biggest fan of, but still code is clearer and I guess
it is the right thing to do.

Also what worries me that we still have *.transform set of functions that duplicates this.
This will be addressed separately