ExportAssignmentDeclaration is not a TopLevelDeclaration because we now have ModuleDeclaration::export property
accordingly ExportAssignmentNode is not a TopLevelNode

Two things important about this commit.
First, there's regression in js-type-analysis which yet to be adressed.
Second, more important - this is first time when proto model and
declaration model are not mirroring each other closely.

Strictly speaking, we losing information when having a separate export
property - because we can not restore the position of the declaration in
the source code. But I can not think of any scenario when we will suffer
from this. Let it be!